Makes a function in kubectl/util to be exportable
This change is required in order to implement https://github.com/kubernetes/kubectl/issues/538

`LookupContainerPortNumberByName` can be reused in the portforward command when we need to
find a port by name. This commit also moves the function into a separate file,
because `service_port.go` doesn't sound like a good location
for a function that operates on a Pod.
