+++ pkg/kubectl/util/service_port_test.go
@@ -23,81 +23,6 @@ import (
 	"k8s.io/apimachinery/pkg/util/intstr"
 )
 
-func TestLookupContainerPortNumberByName(t *testing.T) {
-	tests := []struct {
-		name     string
-		pod      v1.Pod
-		portname string
-		portnum  int32
-		err      bool
-	}{
-		{
-			name: "test success 1",
-			pod: v1.Pod{
-				Spec: v1.PodSpec{
-					Containers: []v1.Container{
-						{
-							Ports: []v1.ContainerPort{
-								{
-									Name:          "https",
-									ContainerPort: int32(443)},
-								{
-									Name:          "http",
-									ContainerPort: int32(80)},
-							},
-						},
-					},
-				},
-			},
-			portname: "http",
-			portnum:  int32(80),
-			err:      false,
-		},
-		{
-			name: "test faulure 1",
-			pod: v1.Pod{
-				Spec: v1.PodSpec{
-					Containers: []v1.Container{
-						{
-							Ports: []v1.ContainerPort{
-								{
-									Name:          "https",
-									ContainerPort: int32(443)},
-							},
-						},
-					},
-				},
-			},
-			portname: "www",
-			portnum:  int32(0),
-			err:      true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			portnum, err := lookupContainerPortNumberByName(tt.pod, tt.portname)
-			if err != nil {
-				if tt.err {
-					return
-				}
-
-				t.Errorf("%v: unexpected error: %v", tt.name, err)
-				return
-			}
-
-			if tt.err {
-				t.Errorf("%v: unexpected success", tt.name)
-				return
-			}
-
-			if portnum != tt.portnum {
-				t.Errorf("%v: expected port number %v; got %v", tt.name, tt.portnum, portnum)
-			}
-		})
-	}
-}
-
 func TestLookupContainerPortNumberByServicePort(t *testing.T) {
 	tests := []struct {
 		name          string
